Some updates to try to support kitchen-terraform 3.1.0.
I never got this completely working (kitchen-terraform was only
executing the root module, rather than apply-all'ing to all the
modules).

This is as far as I got before ultimately deciding that
kitchen-terraform wasn't adding enough value given the number of hacks
it took to get it working with terragrunt *-all.

(The hard-coded /usr/local/bin/terraform as part of the workaround
for kitchen-terraform dropping driver:cli; it is not particularly
portable.)