GPII-3103: Check versions of key tools.
This is kind of crude, and duplicates information in both aws/README.md
and ansible-gpii-ci-worker.

However, there are often painful consequences for accidentally using the
wrong version of some key tools, e.g. a new version of terraform
upgrading remote state and making it incompatible with earlier versions,
or a new version of kops triggering an unexpected rolling-update.