GPII-3832: Two variables for port number as used by LFM and CBFM
Updated to reflect universal codebase where two environment variables
define the port number individually for each of the local
flowmanager (LFM) and the cloud-based flowmanager (CBFM).  The LFM is
instantiated by and runs within the "productiontests" container.  In
contrast, the CBFM is already running as a service within the cluster
with a port number defined by its deployment specification.